move time display to a client component to show client time
